requirements.txt
PyQt5 == 5.15.4
Pillow == 8.4.0
pygame == 2.1.2